ai-models/requirements-ai.txt
# HealthFit AI - AI Models Requirements
# Install with: pip install -r requirements-ai.txt

# Web framework for AI API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# Deep Learning (TensorFlow)
tensorflow==2.15.0
keras==2.15.0

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
Pillow==10.2.0

# Data processing
scipy==1.12.0
matplotlib==3.8.2

# Environment
python-dotenv==1.0.0

# HTTP
requests==2.31.0
httpx==0.26.0
